Move ingredient controllers in own subdirectory
